pyproject.toml
[tool.poetry]
name = "gfmodules-national-referral-index"
version = "0.1.0"
description = ""
license = "EUPL-1.2"
authors = ["Ministerie van Volksgezondheid, Welzijn en Sport"]
repository = "https://github.com/minvws/gfmodules-national-referral-index"
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.26.0"
fastapi = "^0.109.1"
mypy = "^1.8.0"
pydantic-settings = "^2.1.0"
inject = "^5.2.0"
sqlalchemy = "^2.0.25"
psycopg = { extras = ["binary", "pool"], version = "^3.1.17" }
types-requests = "^2.31.0.20240406"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.24.0"
opentelemetry-instrumentation = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-requests = "^0.45b0"
requests = "^2.32.0"
statsd = "^4.0.1"
puzi = {git = "https://github.com/minvws/puzi-python" }


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
ruff = "^0.1.13"
safety = "^2.3.5"
codespell = "^2.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
cache-dir = "~/.cache/ruff"
include = [
    "pyproject.toml",
    "app/*.py",
    "tests/*.py",
]

[tool.mypy]
files = "app,tests"
python_version = "3.11"
strict = true
cache_dir = "~/.cache/mypy"
ignore_missing_imports = true

[tool.pytest.ini_options]
cache_dir = "~/.cache/pytest"
testpaths = [
    "tests",
]
